% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{calculateDescriptives}
\alias{calculateDescriptives}
\title{Calculate descriptive statistics.}
\usage{
calculateDescriptives(
  GADSdat.obj,
  varinfo,
  verbose = TRUE,
  showCallOnly = FALSE
)
}
\arguments{
\item{GADSdat.obj}{Object of class \code{GADSdat}, created for example by \code{import_spss} from the \code{eatGADS} package.}

\item{varinfo}{\code{data.frame} with variable information}

\item{verbose}{Cat to console?}

\item{showCallOnly}{tbd}
}
\value{
Returns a list of descriptive statistics.
}
\description{
Calculate descriptive statistics which should be included in the codebook.
}
\examples{
# tbd

}
