% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkScaleConsistency.r
\name{checkScaleConsistency}
\alias{checkScaleConsistency}
\title{Check whether the mean of a scale variable equals the pooled mean of single items}
\usage{
checkScaleConsistency(
  GADSdat,
  inputForDescriptives,
  id,
  tolerance = 0.02,
  verbose = TRUE
)
}
\arguments{
\item{GADSdat}{Object of class\code{GADSdat}}

\item{inputForDescriptives}{The output of the function \code{createInputForDescriptives()}}

\item{id}{Name or column number of the ID variable. Argument can be numeric or character}

\item{tolerance}{A positive numeric value, indicating the maximum allowed discrepancy between the mean of a scale variable and the pooled mean of single items}

\item{verbose}{Logical: Print informations on console?}
}
\value{
Function does not return output but provide messages.
}
\description{
Function need the output of the \code{createInputForDescriptives()} function.
For each non-imputed scale, the function checks whether the (unweighted) mean
of the scale variable equals the pooled mean of the single items which form the scale.
}
