% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create codebook script.}
\usage{
codebook(
  varInfo,
  missings,
  struc,
  scaleInfo,
  register = NULL,
  dat,
  Kennwertedatensatz,
  chapters,
  deckblatt,
  intro,
  literatur,
  abkuerzverz,
  hintmod,
  lastpage
)
}
\arguments{
\item{varInfo}{\code{data.frame} or list of \code{data.frames} containing the variable information, imported via \code{\link{getVarInfo}}.}

\item{missings}{\code{data.frame} or list of \code{data.frames} containing the missing information, imported via \code{\link{getMissings}}.}

\item{struc}{\code{data.frame} containing the structure of the codebook, imported via \code{\link{getStructure}}.}

\item{scaleInfo}{\code{data.frame} containing the information on scales, imported via \code{\link{getScaleInfo}}.}

\item{register}{\code{data.frame} containing the information on the register, imported via \code{\link{getRegister}}.
If \code{NULL}, now register is created. If there are registers for some data sets but not all, the missing registers are simply omitted.}

\item{dat}{\code{data.frame} or list of \code{data.frames} containing the data sets, imported via \code{\link{import_spss}}.}

\item{Kennwertedatensatz}{\code{data.frame} or list of \code{data.frame} containing the descriptive statistics, imported via \code{\link{calculateDescriptives}}.}

\item{chapters}{\code{data.frame} or list of \code{data.frames} containing the chapter information, imported via \code{\link{getChapters}}.
Determines the order of chapters in the codebook.}

\item{deckblatt}{Character vector with the cover page, created via \code{\link{makeCover}}.}

\item{intro}{Character vector, introduction.}

\item{literatur}{Character vector with the literature information, created via \code{\link{makeLit}}.}

\item{abkuerzverz}{Character vector with the abbreviations, created via \code{\link{makeAbbrList}}.}

\item{hintmod}{Character vector with the information on the background model, created via \code{\link{makeBGM}}.}

\item{lastpage}{Character vektor, last page.}
}
\value{
Codebook latex script.
}
\description{
Create the complete codebook latex script.
}
\examples{
#tbd

}
