% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAbbrList.R
\name{createAbbrList}
\alias{createAbbrList}
\title{Create abbreviation list template.}
\usage{
createAbbrList(
  headings = list(Akronyme = c("Abkuerzung", "Bedeutung"), `Statistische Formelzeichen` =
    c("Symbol", "Bedeutung"))
)
}
\arguments{
\item{headings}{Headings.}
}
\value{
Returns a latex snippet.
}
\description{
Create abbreviation list template.
}
\examples{
#tbd

}
