% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChapters.R
\name{createChapters}
\alias{createChapters}
\title{Create chapter template.}
\usage{
createChapters(varInfo)
}
\arguments{
\item{varInfo}{Object (either \code{list} or \code{data.frame}) containing variable information.}
}
\value{
Chapter template.
}
\description{
Create chapter template based on variable information. The order of the entries determines the order of the documentation in the codebook.
}
\examples{
#tbd

}
