% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputForDescriptives.r
\name{createInputForDescriptives}
\alias{createInputForDescriptives}
\title{Prepare variable information.}
\usage{
createInputForDescriptives(
  GADSdat.obj,
  idExpr = "^ID",
  impExpr = c("IMPUTATION[[:digit:]]{1,2}$", "PV[[:digit:]]{1,2}"),
  scaleExpr = "^Skala",
  varNameSeparatorImp = "_",
  lastOccurrence = TRUE,
  groupSuffixImp = "imp",
  nCatsForScaleDefLowerBound = 2,
  nCatsForScaleDefUpperBound = 5,
  verbose = TRUE
)
}
\arguments{
\item{GADSdat.obj}{Object of class \code{GADSdat}, created by \code{import_spss} from the \code{eatGADS} package, for example. Alternatively, a list of objects of class \code{GADSdat}}

\item{idExpr}{Regular expression to identify ID variables from variable names (Note: for multiple expressions, i.e. if \code{idExpr} is a character vector of length > 1, at least one expression should match to identify the variable as ID variable)}

\item{impExpr}{Regular expression to identify imputed variables from variable labels in GADSdat object (Note: for multiple expressions, i.e. if \code{impExpr} is a character vector of length > 1, at least one expression should match to identify the variable as an imputed variable)}

\item{scaleExpr}{Regular expression to identify scale variables from variable labels in GADSdat object (Note: for multiple expressions, i.e. if \code{scaleExpr} is a character vector of length > 1, at least one expression should match to identify the variable as a scale variable)}

\item{varNameSeparatorImp}{character sign to separate the "pooled" suffix from group name in group column}

\item{lastOccurrence}{Logical: If varNameSeparatorImp occurrs multiple times within a string, lastOccurrence defines whether the last occurrence should be used for splitting}

\item{groupSuffixImp}{tbd}

\item{nCatsForScaleDefLowerBound}{Lower bound for scale definition of categorical variables: Variables with this number of categories are defined to be nominal (instead of ordinal)}

\item{nCatsForScaleDefUpperBound}{Upper bound for scale definition of categorical variables: Variables with more categories than defined in upper bound are defined to be nominal (instead of ordinal)}

\item{verbose}{tbd}
}
\value{
Returns a \code{data.frame} with variable information with following columns
\itemize{
 \item \code{varName} The name of the variable as it occurs in the data
 \item \code{varLabel} The label of the variable as it occurs in the \code{GADSdat} label sheet
 \item \code{format} The variable format as displayed in the labels sheet of the \code{GADSdat} object
 \item \code{imp} Logical: Whether or not the variable is imputed
 \item \code{type} The type of the variable. Two possible entries, \code{variable} or \code{scale}
 \item \code{scale} The scale level of the variable. Possible entries: \code{nominal}, \code{ordinal}, \code{numeric}. ID variables and character variables have missing entries in this column
 \item \code{group} If the variable is part of a scale with several items, a common entry in the group column indicates that these variables belong together
}
}
\description{
Create a variable information data.frame from the GADSdat object
}
\examples{
varInfo <- createInputForDescriptives(eatGADS::pisa, impExpr = "Plausible Value")

}
