% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetadata.R
\name{createMetadata}
\alias{createMetadata}
\title{Create meta data template.}
\usage{
createMetadata()
}
\value{
Returns a meta data template.
}
\description{
Create meta data template.
}
\examples{
createMetadata

}
