% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMissings.R
\name{createMissings}
\alias{createMissings}
\title{Create missing value information.}
\usage{
createMissings(GADSdat, inputForDescriptives)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{inputForDescriptives}{Input for descriptive statistics calculation.}
}
\value{
Returns the value information.
}
\description{
Create information on value level (value labels and missing codes).
}
\examples{
#tbd

}
