% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegister.R
\name{createRegister}
\alias{createRegister}
\title{Create register template.}
\usage{
createRegister(varinfo, keywordList)
}
\arguments{
\item{varinfo}{Object (either list or data.frame) containing variable information.}

\item{keywordList}{Character vector of keyword columns to be added.}
}
\value{
Register template.
}
\description{
Create register template based on variable information.
}
\examples{
#tbd

}
