% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createScaleInfo.R
\name{createScaleInfo}
\alias{createScaleInfo}
\title{Create scale information.}
\usage{
createScaleInfo(varInfo)
}
\arguments{
\item{varInfo}{\code{varInfo} object.}
}
\value{
Returns the scale information.
}
\description{
Create information on scale aggregation.
}
\examples{
#tbd

}
