% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructure.R
\name{createStructure}
\alias{createStructure}
\title{Create structure template.}
\usage{
createStructure(varInfo = NULL)
}
\arguments{
\item{varInfo}{Object (either list or data.frame) containing variable information.}
}
\value{
Structure template.
}
\description{
Create structure template based on variable information.
}
\details{
To create the structure template, information from the columns \code{'Unterteilung.im.Skalenhandbuch'} and \code{'Gliederung'} are used.
}
\examples{
#tbd

}
