% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarInfo.R
\name{createVarInfo}
\alias{createVarInfo}
\title{Create variable information template.}
\usage{
createVarInfo(
  GADSdat,
  inputForDescriptives,
  encodingList = NULL,
  makeStructure = FALSE
)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}

\item{inputForDescriptives}{Input for descriptives calculation.}

\item{encodingList}{tbd.}

\item{makeStructure}{Should an automatic structuring of variables be created}
}
\value{
Returns the variable information template.
}
\description{
Create variable information (background model, which variables in the data set, ...) template.
}
\examples{
#tbd

}
