% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVarueInfo.R
\name{createVarueInfo}
\alias{createVarueInfo}
\title{Create variable information template.}
\usage{
createVarueInfo(
  GADSdat,
  inputForDescriptives,
  encodingList = NULL,
  makeStructure = TRUE
)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}

\item{inputForDescriptives}{Input for descriptives calculation.}

\item{encodingList}{tbd.}

\item{makeStructure}{Should an automatic structuring of variables be created_}
}
\value{
Returns the variable information template.
}
\description{
Create variable information (background model, which variables in the data set, ...) template.
}
\examples{
#tbd

}
