% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedLatexSnippet.R
\name{embedLatexSnippet}
\alias{embedLatexSnippet}
\title{Embed a Latex Snippet.}
\usage{
embedLatexSnippet(snippet)
}
\arguments{
\item{snippet}{A Latex snippet.}
}
\value{
A testable latex script.
}
\description{
Embed a Latex Snippet in the Latex environment used by \code{\link{codebook}}..
}
\examples{
latexSnippet <- embedLatexSnippet("This is a sentence.")

}
