% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbbrList.R
\name{getAbbrList}
\alias{getAbbrList}
\title{Import abbreviation list.}
\usage{
getAbbrList(filePath)
}
\arguments{
\item{filePath}{Path to excel file.}
}
\value{
Abbreviation list.
}
\description{
Import abbreviation list.
}
\examples{
#tbd

}
