% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChapters.R
\name{getChapters}
\alias{getChapters}
\title{Import chapter structure information.}
\usage{
getChapters(filePath)
}
\arguments{
\item{filePath}{Path to excel file.}
}
\value{
Chapter structure information.
}
\description{
Import chapter structure information.
}
\examples{
#tbd

}
