% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExcel.R
\name{getExcel}
\alias{getExcel}
\title{Import data from Excel.}
\usage{
getExcel(filePath)
}
\arguments{
\item{filePath}{Path to excel file.}
}
\value{
Either a \code{data.frame} (if the Excel file has a single sheet)
or a named \code{list} (if the Excel file has multiple sheets).
}
\description{
Import all sheets from an Excel file.
}
\examples{
#tbd

}
