% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInputForDescriptives.R
\name{getInputForDescriptives}
\alias{getInputForDescriptives}
\title{Import input for descriptives information.}
\usage{
getInputForDescriptives(filePath)
}
\arguments{
\item{filePath}{Path to excel file.}
}
\value{
variable information.
}
\description{
Import input for descriptives information.
}
\examples{
#tbd

}
