% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLitInfo.R
\name{getLitinfo}
\alias{getLitinfo}
\title{Import register excel.}
\usage{
getLitinfo(filePath)
}
\arguments{
\item{filePath}{Path to excel file.}
}
\value{
Register.
}
\description{
Import register excel.
}
\examples{
#tbd

}
