% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMissings.R
\name{getMissings}
\alias{getMissings}
\title{Get value information.}
\usage{
getMissings(filePath)
}
\arguments{
\item{filePath}{Path to the excel-file.}
}
\value{
Returns the value information.
}
\description{
Get information on value level (value labels and missing codes).
}
\examples{
tbd

}
