% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegister.R
\name{getRegister}
\alias{getRegister}
\title{Import register excel.}
\usage{
getRegister(filePath)
}
\arguments{
\item{filePath}{Path to excel file.}
}
\value{
Register.
}
\description{
Import register excel.
}
\examples{
tbd

}
