% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScaleInfo.R
\name{getScaleInfo}
\alias{getScaleInfo}
\title{Get scale information.}
\usage{
getScaleInfo(filePath)
}
\arguments{
\item{filePath}{Path to the excel-file.}
}
\value{
Returns the scale information.
}
\description{
Get information on scale aggregation.
}
\examples{
#tbd

}
