% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructure.R
\name{getStructure}
\alias{getStructure}
\title{Import structure information.}
\usage{
getStructure(filePath)
}
\arguments{
\item{filePath}{Path to excel file.}
}
\value{
Structure information.
}
\description{
Import structure information.
}
\examples{
#tbd

}
