% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarInfo.R
\name{getVarInfo}
\alias{getVarInfo}
\title{Import variable information.}
\usage{
getVarInfo(filePath)
}
\arguments{
\item{filePath}{Path to excel file.}
}
\value{
variable information.
}
\description{
Import variable information.
}
\examples{
#tbd

}
