% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferLayout.R
\name{inferLayout}
\alias{inferLayout}
\title{Infer layout column in a variable information.}
\usage{
inferLayout(varInfo, GADSdat, inputForDescriptives)
}
\arguments{
\item{varInfo}{\code{varInfo} object.}

\item{GADSdat}{\code{GADSdat} object.}

\item{inputForDescriptives}{Input for descriptive statistics calculation.}
}
\value{
Returns the modified variable information.
}
\description{
Temporary function: Infers the layout numeric code from the \code{SPSS} format of a variable in the data and the information
in the \code{inputForDescriptives}.
}
\examples{
#tbd

}
