% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{kds.erstellen}
\alias{kds.erstellen}
\title{Calculate descriptive statistics.}
\usage{
kds.erstellen(
  fbshort,
  varue.info,
  varue.missings,
  skalen.info,
  ds,
  variablen.,
  id
)
}
\arguments{
\item{fbshort}{Data set name.}

\item{varue.info}{Varue info.}

\item{varue.missings}{Missing codes.}

\item{skalen.info}{Scale informations.}

\item{ds}{Data set.}

\item{variablen.}{further arguments passed to or from other methods.}

\item{id}{ID variable in ds.}
}
\value{
Returns a list of descriptive statistics.
}
\description{
Calculate descriptive statistics which should be included in the codebook.
}
\examples{
# tbd

}
