% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAbbrList.R
\name{makeAbbrList}
\alias{makeAbbrList}
\title{Create abbreviation list.}
\usage{
makeAbbrList(
  filePath,
  captions = list(Akronyme =
    "Abk{\\\\\\"u}rzungen", `Statistische Formelzeichen` = "Statistische Formelzeichen"),
  sort.entries = TRUE
)
}
\arguments{
\item{filePath}{Path to the file.}

\item{captions}{tbd}

\item{sort.entries}{Should the abbreviation lists be sorted alphabetically?}
}
\value{
Returns a latex snippet.
}
\description{
Create abbreviation lists latex snippet.
}
\examples{
#tbd

}
