% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAbbrList.R
\name{makeAbkVerz}
\alias{makeAbkVerz}
\title{Create abbreviation list.}
\usage{
makeAbkVerz(
  varue.file,
  sheets = c("Akronyme", "Statistische Formelzeichen"),
  headings = list(Akronyme = c("Abkürzung", "Bedeutung"), `Statistische Formelzeichen`
    = c("Symbol", "Bedeutung")),
  captions = list(Akronyme = c("Abkürzungen"), `Statistische Formelzeichen` =
    c("Statistische Formelzeichen")),
  sort.entries = c(TRUE, TRUE)
)
}
\arguments{
\item{logoFile}{Path to the graphic file.}

\item{maintitle}{Main title of the document.}

\item{subtitle}{Sub title of the document.}

\item{authors}{Authors of the document.}

\item{addAuthors}{Additional contributers to the document.}

\item{schriftenreihe}{Schriftenreihe of the document.}

\item{bibinfo}{Bibiolografic info of the document.}
}
\value{
Returns a latex snippet.
}
\description{
Create abbreviation lists.
}
\examples{
tbd

}
