% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCover.R
\name{makeCover}
\alias{makeCover}
\title{Create cover page.}
\usage{
makeCover(
  logoFile = NULL,
  maintitle = NULL,
  subtitle = NULL,
  authors = NULL,
  addAuthors = NULL,
  schriftenreihe = NULL,
  bibinfo = NULL
)
}
\arguments{
\item{logoFile}{Path to the graphic file.}

\item{maintitle}{Main title of the document.}

\item{subtitle}{Sub title of the document.}

\item{authors}{Authors of the document.}

\item{addAuthors}{Additional contributers to the document.}

\item{schriftenreihe}{Schriftenreihe of the document.}

\item{bibinfo}{Bibiolografic info of the document.}
}
\value{
Returns a latex snippet.
}
\description{
Create a cover page.
}
\examples{
makeCover(maintitle = "Study of Achievement", subtitle = "Codebook of Study of Achievement",
         authors = "Some Person", addAuthors = "With the help of some other persons",
         schriftenreihe = "Book 9 of Studies of Achievement")

}
