% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLit.R
\name{makeLit}
\alias{makeLit}
\title{Create literature latex snippet.}
\usage{
makeLit(litInfo)
}
\arguments{
\item{litInfo}{Literature information..}
}
\value{
Literature latex snippet.
}
\description{
Create literature latex snippet.
}
\examples{
#tbd

}
