% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMetadata.R
\name{makeMetadata}
\alias{makeMetadata}
\title{Create meta data snippet.}
\usage{
makeMetadata(filePath)
}
\arguments{
\item{filePath}{Path to file.}
}
\value{
Returns a latex snippet.
}
\description{
Create meta data latex snippet.
}
\examples{
#tbd

}
