% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeExcel.R
\name{writeExcel}
\alias{writeExcel}
\title{Save to \code{.xlsx}.}
\usage{
writeExcel(df_list, filePath, row.names = FALSE, col.names = TRUE)
}
\arguments{
\item{df_list}{A named list with \code{data.frames} to be written to \code{.xlsx}. The names of the list become the sheet names.}

\item{filePath}{The path to the output file.}

\item{row.names}{Logical: Should row names of the \code{data.frames} be written to file?}

\item{col.names}{Logical: Should column names of the \code{data.frames} be written to file?}
}
\description{
Create an \code{.xlsx} file with multiple sheets.
}
\details{
An existing file is overwritten.
}
\examples{
f <- tempfile(fileext = ".xlsx")
writeExcel(mtcars, filePath = f)

}
