% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDescriptives.R
\name{calculateDescriptives}
\alias{calculateDescriptives}
\title{Calculate descriptive statistics.}
\usage{
calculateDescriptives(
  GADSdat,
  inputForDescriptives,
  verbose = TRUE,
  showCallOnly = FALSE
)
}
\arguments{
\item{GADSdat}{Object of class \code{GADSdat}, created for example by \code{import_spss} from the \code{eatGADS} package.}

\item{inputForDescriptives}{\code{data.frame} with variable information. This table can be created from GADSdat object, using the \code{createInputForDescriptives} function}

\item{verbose}{Cat to console?}

\item{showCallOnly}{Logical: only for diagnostics. If TRUE, no calculation is proceed, and only the function which is called for calculation is returned.}
}
\value{
Returns a list of descriptive statistics.
}
\description{
Calculate descriptive statistics which should be included in the codebook.
}
\examples{
# tbd

}
