% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create codebook script.}
\usage{
codebook(
  varue.info,
  varue.missings,
  varue.gliederung,
  skalen.info,
  varue.reg,
  make.reg = NULL,
  Gesamtdatensatz,
  Kennwertedatensatz,
  variablen,
  id,
  fbshort,
  fblong,
  deckblatt,
  intro,
  literatur,
  abkuerzverz,
  hintmod,
  lastpage
)
}
\arguments{
\item{varue.info}{tbd}

\item{varue.missings}{Liste mit data.frames der Uebersichten der Werteinformationen}

\item{varue.gliederung}{Liste mit data.frames der Uebersichten der Gliederungsinformationen}

\item{skalen.info}{data.frame, Skaleninformationen ueber alle Fragebogen}

\item{varue.reg}{Liste mit data.frames der Uebersichten der Registerinformationen}

\item{make.reg}{tbd}

\item{Gesamtdatensatz}{Liste mit data.frames der Datensaetze}

\item{Kennwertedatensatz}{Liste mit data.frames der Kennwertedatensaetze}

\item{variablen}{Liste mit character-Vektoren der zu berichtenden Variablen}

\item{id}{tbd}

\item{fbshort}{Character-Vektor, Fragebogenkuerzel}

\item{fblong}{Character-Vektor, Namen der Fragebogen, wie sie im Skalenhandbuch ausformuliert genannt werden}

\item{deckblatt}{tbd}

\item{intro}{Character-Vektor, Einleitung}

\item{literatur}{Character-Vektor, Literaturverzeichnis}

\item{abkuerzverz}{Character-Vekotr, Tabelle Abkuerzungsverzeichnis}

\item{hintmod}{Character-Vektor, Tabelle Hintergrundmodell}

\item{lastpage}{Character-Vektor, Letzte Seite}
}
\value{
Codebook latex script.
}
\description{
Create the complete codebook latex script.
}
\examples{
#tbd

}
