% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create codebook script.}
\usage{
codebook(
  varInfo,
  missings,
  struc,
  scaleInfo,
  register,
  make.reg = NULL,
  dat,
  Kennwertedatensatz,
  id,
  fbshort,
  fblong,
  deckblatt,
  intro,
  literatur,
  abkuerzverz,
  hintmod,
  lastpage
)
}
\arguments{
\item{varInfo}{\code{data.frame} or list of \code{data.frames} containing the variable information, imported via \code{\link{getVarInfo}}.}

\item{missings}{\code{data.frame} or list of \code{data.frames} containing the missing information, imported via \code{\link{getMissings}}.}

\item{struc}{\code{data.frame} containing the structure of the codebook, imported via \code{\link{getStructure}}.}

\item{scaleInfo}{\code{data.frame} containing the information on scales, imported via \code{\link{getScaleInfo}}.}

\item{register}{\code{data.frame} containing the information on the register, imported via \code{\link{getRegister}}.}

\item{make.reg}{Logical. Should a register be created?}

\item{dat}{\code{data.frame} or list of \code{data.frames} containing the data sets, imported via \code{\link{eatGADS::import_spss}}.}

\item{Kennwertedatensatz}{\code{data.frame} or list of \code{data.frame} containing the descriptive statistics, imported via \code{\link{calculateDescriptives}}.}

\item{id}{Named character vector containing the identifier variable in each data set.}

\item{fbshort}{Character vector with the short names of the data sets.}

\item{fblong}{Character vector with the long names of the data sets as they should appear in the codebook.}

\item{deckblatt}{Character vector with the cover page, created via \code{\link{makeCover}}.}

\item{intro}{Character vector, introduction.}

\item{literatur}{Character vector with the literature information, created via \code{\link{makeLit}}.}

\item{abkuerzverz}{Character vector with the abbreviations, created via \code{\link{makeAbbrList}}.}

\item{hintmod}{Character vector with the information on the background model, created via \code{\link{makeBGM}}.}

\item{lastpage}{Character vektor, last page.}
}
\value{
Codebook latex script.
}
\description{
Create the complete codebook latex script.
}
\examples{
#tbd

}
