% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputForDescriptives.r
\name{createInputForDescriptives}
\alias{createInputForDescriptives}
\title{Create input data.frame for subsequent calculation of descriptives}
\usage{
createInputForDescriptives(
  GADSdat.obj,
  idExpr = "^ID",
  impExpr = c("IMPUTATION[[:digit:]]{1,2}$", "PV[[:digit:]]{1,2}"),
  scaleExpr = "^Skala",
  varNameSeparatorImp = "_",
  ncharSeparatorImp = 1,
  lastOccurrence = TRUE,
  groupSuffixImp = "imp",
  nCatsForOrdinal = c(2:5),
  verbose = TRUE
)
}
\arguments{
\item{GADSdat.obj}{Object of class \code{GADSdat}, created by \code{import_spss} from the \code{eatGADS} package, for example. Alternatively, a list of objects of class \code{GADSdat}}

\item{idExpr}{Regular expression to identify ID variables from variable names (Note: for multiple expressions, i.e. if \code{idExpr} is a character vector of length > 1, at least one expression should match to identify the variable as ID variable)}

\item{impExpr}{Regular expression to identify imputed variables from variable labels in GADSdat object (Note: for multiple expressions, i.e. if \code{impExpr} is a character vector of length > 1, at least one expression should match to identify the variable as an imputed variable)}

\item{scaleExpr}{Regular expression to identify scale variables from variable labels in GADSdat object (Note: for multiple expressions, i.e. if \code{scaleExpr} is a character vector of length > 1, at least one expression should match to identify the variable as a scale variable)}

\item{varNameSeparatorImp}{character sign to separate the "pooled" suffix from group name in group column. If no such sign exists in the data, i.e. if multiple imputations occur as \code{pv1}, \code{pv2}, \code{pv3}, instead of \code{pv_1}, \code{pv_2}, \code{pv_3}, or \code{pv.1}, \code{pv.2}, \code{pv.3}, use \code{NA} or \code{NULL} or \code{""}.}

\item{ncharSeparatorImp}{Integer: only relevant if no \code{varNameSeparatorImp} exists, i.e. if multiple imputations occur as \code{pv1}, \code{pv2}, \code{pv3}, instead of \code{pv_1}, \code{pv_2}, \code{pv_3}, or \code{pv.1}, \code{pv.2}, \code{pv.3}. \code{ncharSeparatorImp} than specifies the number of character signs which should be dropped to identify the common variable stem. If \code{varNameSeparatorImp} is not \code{NA} or \code{NULL} or \code{""}, \code{ncharSeparatorImp} will be ignored. For example, if multiple imputations occur as \code{pv_1}, \code{pv_2}, \code{pv_3}, use \code{varNameSeparatorImp = "_"}. If multiple imputations occur as \code{pv.1}, \code{pv.2}, \code{pv.3}, use \code{varNameSeparatorImp = "\\."}. If multiple imputations occur as \code{pv1}, \code{pv2}, \code{pv3}, use \code{varNameSeparatorImp = NULL} and \code{ncharSeparatorImp = 1}.}

\item{lastOccurrence}{Logical: If varNameSeparatorImp occurrs multiple times within a string, \code{lastOccurrence} defines whether the last occurrence should be used for splitting}

\item{groupSuffixImp}{tbd}

\item{nCatsForOrdinal}{Numeric vector with number of categories considered for ordinal variables. Variables with number of categories as defined here are considered to be ordinal instead of nominal. If NULL, this rule will be ignored, and nominal/ordinal assignment is done in other ways}

\item{verbose}{tbd}
}
\value{
Returns a \code{data.frame} with variable information with following columns
\itemize{
 \item \code{varName} The name of the variable as it occurs in the data
 \item \code{varLabel} The label of the variable as it occurs in the \code{GADSdat} label sheet
 \item \code{format} The variable format as displayed in the labels sheet of the \code{GADSdat} object
 \item \code{imp} Logical: Whether or not the variable is imputed
 \item \code{type} The type of the variable. Two possible entries, \code{variable} or \code{scale}
 \item \code{scale} The scale level of the variable. Possible entries: \code{nominal}, \code{ordinal}, \code{numeric}. ID variables and character variables have missing entries in this column. Be cautious that 'ordinal' sometimes may be allocated erroneously. The resulting table should be exported to Excel for further checks.
 \item \code{group} If the variable is part of a scale with several items, a common entry in the group column indicates that these variables belong together
}
}
\description{
Create a variable information data.frame from the GADSdat object. This input can be used to calculate the descriptives of the data via the \code{calculateDescriptives} function
}
\examples{
varInfo <- createInputForDescriptives(eatGADS::pisa, impExpr = "Plausible Value")

}
