% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLitInfo.R
\name{createLitInfo}
\alias{createLitInfo}
\title{Create literature template.}
\usage{
createLitInfo(varInfo = NULL)
}
\arguments{
\item{varInfo}{A \code{varInfo} object as imported by \code{getVarInfo}.}
}
\value{
Literature template.
}
\description{
Create literature template.
}
\examples{
#tbd

}
