% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegister.R
\name{createRegister}
\alias{createRegister}
\title{Create register template.}
\usage{
createRegister(inputForDescriptives, keywordList = NULL)
}
\arguments{
\item{inputForDescriptives}{Object (either \code{list} or \code{data.frame}) containing variable information.}

\item{keywordList}{Character vector of keyword columns to be added. If \code{NULL}, no additional columns are added.
If \code{inputForDescriptives} is a \code{list}, \code{keywordList} must also be a \code{list} (of character vectors).}
}
\value{
Register template.
}
\description{
Create register template based on variable information.
}
\examples{
#tbd

}
