% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createScaleInfo.R
\name{createScaleInfo}
\alias{createScaleInfo}
\title{Create scale information.}
\usage{
createScaleInfo(inputForDescriptives)
}
\arguments{
\item{inputForDescriptives}{\code{inputForDescriptives} object as created by the \code{createInputForDescriptives} function}
}
\value{
Returns the scale information.
}
\description{
Create information on scale aggregation. Scales, 'fake scales', and imputed variables are listed.
}
\examples{
#tbd

}
