% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStructure.R
\name{createStructure}
\alias{createStructure}
\title{Create structure template.}
\usage{
createStructure(inputForDescriptives)
}
\arguments{
\item{inputForDescriptives}{Object (either list or data.frame) containing variable information.}
}
\value{
Structure template.
}
\description{
Create structure template based on variable information.
}
\examples{
#tbd

}
