% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExcel.R
\name{getExcel}
\alias{getExcel}
\title{Import data from Excel.}
\usage{
getExcel(filePath, funList = NULL)
}
\arguments{
\item{filePath}{Path to excel file.}

\item{funList}{tbd.}
}
\value{
Either a \code{data.frame} (if the Excel file has a single sheet)
or a named \code{list} (if the Excel file has multiple sheets).
}
\description{
Import all sheets from an Excel file.
}
\examples{
filePath <- system.file("extdata", "example_abbrList.xlsx", package = "eatCodebook")
excel_list   <- getExcel(filePath)

}
