% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLitInfo.R
\name{getLitInfo}
\alias{getLitInfo}
\title{Import literature information.}
\usage{
getLitInfo(filePath)
}
\arguments{
\item{filePath}{Path to excel file.}
}
\value{
Literature information.
}
\description{
Import literature information.
}
\examples{
#tbd

}
