% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAbbrList.R
\name{makeAbbrList}
\alias{makeAbbrList}
\title{Create abbreviation list.}
\usage{
makeAbbrList(
  filePath,
  sheets = c("Akronyme", "Statistische Formelzeichen"),
  headings = list(Akronyme = c("Abkürzung", "Bedeutung"), `Statistische Formelzeichen`
    = c("Symbol", "Bedeutung")),
  captions = list(Akronyme = c("Abkürzungen"), `Statistische Formelzeichen` =
    c("Statistische Formelzeichen")),
  sort.entries = c(TRUE, TRUE)
)
}
\arguments{
\item{filePath}{Path to the file.}

\item{sheets}{Main title of the document.}

\item{headings}{Sub title of the document.}

\item{captions}{Authors of the document.}

\item{sort.entries}{Additional contributers to the document.}
}
\value{
Returns a latex snippet.
}
\description{
Create abbreviation lists.
}
\examples{
#tbd

}
