% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBGM.R
\name{makeBGM}
\alias{makeBGM}
\title{Create background model information.}
\usage{
makeBGM(varueInfo)
}
\arguments{
\item{varueInfo}{Variable information}
}
\value{
Returns a latex snippet.
}
\description{
Create background model information.
}
\examples{
#tbd

}
