% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRegister.R
\name{makeRegister}
\alias{makeRegister}
\title{Create register.}
\usage{
makeRegister(fblong, fb.akt, varue.reg, double.vars)
}
\arguments{
\item{fblong}{Full name of data set}

\item{fb.akt}{Fragebogenkuerzel aus fbshort}

\item{varue.reg}{Informationen zum Register: data.frame, Spalten sind Schlagwoerter, die im Register aufgelistet sind,
Zeilen sind VarNamen. Eintraege sind \code{"x"} oder \code{""}, ob Variable unter dem Schlagwort im Register aufgelistet werden soll.}

\item{double.vars}{tbd}
}
\value{
Returns a (character vector) latex snippet.
}
\description{
Create register latex snippet.
}
\examples{
#tbd

}
