% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVarinfo.r
\name{prepareVarinfo}
\alias{prepareVarinfo}
\title{Prepare variable information.}
\usage{
prepareVarinfo(
  GADSdat.obj,
  varsToExclude = NULL,
  impExpr = c("IMPUTATION[[:digit:]]{1,2}$", "PV[[:digit:]]{1,2}"),
  scaleExpr = "^Skala",
  varNameSeparatorImp = "_",
  lastOccurrence = TRUE,
  groupSuffixImp = "imp",
  verbose = TRUE
)
}
\arguments{
\item{GADSdat.obj}{Object of class GADSdat, created by \code{import_spss} from the \code{eatGADS} package, for example}

\item{varsToExclude}{tbd}

\item{impExpr}{Cat to console?}

\item{scaleExpr}{Cat to console?}

\item{varNameSeparatorImp}{tbd}

\item{groupSuffixImp}{tbd}

\item{verbose}{tbd}

\item{lastOccurence}{tbd}
}
\value{
Returns a \code{data.frame} with variable information.
}
\description{
Create a variable information data.frame.
}
\examples{
varInfo <- prepareVarinfo(eatGADS::pisa, impExpr = "Plausible Value")

}
